Fix ineffective parameter types found on LGTM.com
In TypeScript, the parameters of a function signature must have a name and may
optionally have a type. A common mistake is to try to omit the name. This means
the type is instead seen as the name. As a result, the parameter type will
default to any since no type was given.